flask_app/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
mysql-connector-python
